==> analysis_easifish/bydatasets/sparse06/00.reformat_spots.ipynb <==
import os, sys
sys.path.append('..')

import pandas as pd
import numpy as np
import zarr
from easi_fish import n5_metadata_utils as n5mu

# compare spots 

ddir  = '/u/home/f/f7xiesnm/project-zipursky/easifish/sparse06/outputs/r1'
imdir = "/u/home/f/f7xiesnm/project-zipursky/easifish/sparse06/outputs/r1/stitching/export.n5"
# ddir  = '/u/home/f/f7xiesnm/project-zipursky/easifish/lt186/outputs/r1'
# imdir = "/u/home/f/f7xiesnm/project-zipursky/easifish/lt186/outputs/r1/stitching/export.n5"

for c in [0,2,4]:
    f1 = ddir+f"/spots/spots_rsfish_c{c}.csv"
    fout = f1.replace('_rsfish', '').replace('csv', 'txt')
    
    input_voxels        = f1 
    output_microns      = fout
    n5_path             = imdir
    extraction_subpath  = '/c0/s0'

    vox    = n5mu.read_voxel_spacing(n5_path, extraction_subpath)
    # vox  = [0.23, 0.23, 0.42]
    grid = n5mu.read_voxel_grid(n5_path, extraction_subpath)
    print(vox, grid)

    rsfish_spots = np.loadtxt(input_voxels, delimiter=',', skiprows=1)
    rsfish_spots[:, :3] = rsfish_spots[:, :3] * vox

    # Remove unnecessary columns (t,c) at indexes 3 and 4 
    rsfish_spots = np.delete(rsfish_spots, np.s_[3:5], axis=1)

    print(f"Saving {rsfish_spots.shape[0]} points in micron space to", output_microns)
    np.savetxt(output_microns, rsfish_spots, delimiter=',')
    
    # break
