==> analysis_easifish/bydatasets/sparse06/05.viz_s0.ipynb <==
# Inputs
---

import numpy as np, zarr, nrrd
# alignment functions
from bigstream.align import alignment_pipeline
from bigstream.transform import apply_transform
import tifffile

import seaborn as sns
import matplotlib.pyplot as plt

p1      = '/u/home/f/f7xiesnm/project-zipursky/data/01_fused'
p2      = '/u/home/f/f7xiesnm/project-zipursky/data/03_flatfused'
outdir = '/u/home/f/f7xiesnm/project-zipursky/easifish/results/viz_sparse06_jan23'
!mkdir -p $outdir

!ls $p1

!ls $p2

# s0

path = p1 + f'/sparse06_r1_autos1_export.n5'
zarr_data = zarr.open(store=zarr.N5Store(path), mode='r')
for c in ['c0']:
    img = zarr_data[f'/{c}/s0'] 
    print(img.shape)
    # print(np.max(img), np.min(img))
    # output = outdir+f'/demo_small_s1-ds-try3_{c}_s3.tiff'
    # print(output)
    # tifffile.imwrite(output, img)

3966/960, 4966/960, 0.13/3, 0.17/4


img_s = img[1000]

plt.imshow(img_s, 
           vmax=np.percentile(img_s, 95), 
           vmin=np.percentile(img_s, 0.1),)

# s0 - fuseflat

path = p2 + f'/sparse06_r1_autos1_flatfused.n5'

zarr_data = zarr.open(store=zarr.N5Store(path), mode='r')
for c in ['c0']:
    img = zarr_data[f'/{c}/s0'] 
    print(img.shape)
    # print(np.max(img), np.min(img))
    # output = outdir+f'/demo_small_s1-ds-try3_{c}_s3.tiff'
    # print(output)
    # tifffile.imwrite(output, img)

3966/960, 4966/960, 0.13/3, 0.17/4


img_s = img[1000]

plt.imshow(img_s, 
           vmax=np.percentile(img_s, 95), 
           vmin=np.percentile(img_s, 0.1),)

# s3


rounds = ['r1']
for r in rounds:
    path = p1 + f'/sparse06_{r}_export.n5'
    zarr_data = zarr.open(store=zarr.N5Store(path), mode='r')
    for c in ['c0']:
        img = zarr_data[f'/{c}/s3'] 
        print(img.shape)
        print(np.max(img), np.min(img))
        output = outdir+f'/{r}_{c}_s3_flat.tiff'
        print(output)
        tifffile.imwrite(output, img)


rounds = ['r1']
for r in rounds:
    path = p2 + f'/sparse06_{r}_flatfused.n5'
    zarr_data = zarr.open(store=zarr.N5Store(path), mode='r')
    for c in ['c0']:
        img = zarr_data[f'/{c}/s3'] 
        print(img.shape)
        print(np.max(img), np.min(img))
        output = outdir+f'/{r}_{c}_s3_flatfused.tiff'
        print(output)
        tifffile.imwrite(output, img)
